refactored code, now have user to brewery relationships for userBreweryCard. Will have to conditionally render reviews and buttons from there.
