+++ api/database.json
@@ -128524,68 +128524,21 @@
       "userId": 1,
       "breweryId": "6553",
       "beenToBrewery": true,
-      "review": "We love Mayday! Whether it's a date night, a family gathering, or dinner out with friends, Mayday is always the perfect destination. They always have a beer that everyone loves and sodas for the kids. The owner is awesome, as well as a great staff. They have different food trucks depending on the night, but there is always something to eat for everyone.",
-      "brewery": {
-        "id": "6553",
-        "name": "Mayday Brewery",
-        "brewery_type": "micro",
-        "street": "521 Old Salem Rd Ste C",
-        "city": "Murfreesboro",
-        "state": "Tennessee",
-        "postal_code": "37129-5305",
-        "website_url": "http://www.maydaybrewery.com",
-        "phone": "6154799722",
-        "created_at": "2018-07-24 01:34:16.675249",
-        "updated_at": "2018-08-11 21:39:43.116431",
-        "country": "United States",
-        "longitude": "",
-        "latitude": ""
-      },
-      "user": {
-        "email": "cody@home.com",
-        "name": "Cody Jones",
-        "id": 1
-      }
+      "review": "We love Mayday! Whether it's a date night, a family gathering, or dinner out with friends, Mayday is always the perfect destination. They always have a beer that everyone loves and sodas for the kids. The owner is awesome, as well as a great staff. They have different food trucks depending on the night, but there is always something to eat for everyone."
     },
     {
       "id": 2,
       "userId": 1,
       "breweryId": "666",
       "beenToBrewery": false,
-      "review": "Awful place.",
-      "brewery": {
-        "id": "666",
-        "name": "Heretic Brewing Company",
-        "brewery_type": "micro",
-        "street": "1052 Horizon Dr Ste B",
-        "city": "Fairfield",
-        "state": "California",
-        "postal_code": "94533-1694",
-        "website_url": "http://www.hereticbrewing.com",
-        "phone": "7073894573",
-        "created_at": "2018-07-24 01:32:56.401363",
-        "updated_at": "2018-08-11 21:36:06.214918",
-        "country": "United States",
-        "longitude": "",
-        "latitude": ""
-      },
-      "user": {
-        "email": "cody@home.com",
-        "name": "Cody Jones",
-        "id": 1
-      }
+      "review": ""
     },
     {
       "userId": 2,
       "breweryId": "6553",
       "beenToBrewery": true,
-      "id": 3
-    },
-    {
-      "userId": 2,
-      "breweryId": "6553",
-      "beenToBrewery": true,
-      "id": 4
+      "id": 4,
+      "review": ""
     }
   ]
 }
