+++ src/components/breweries/BreweryDetailCardList.js
@@ -1,0 +1,91 @@
+import React, { useState, useEffect } from 'react';
+import { getBreweryById } from '../modules/BreweryManager'
+import { getUserBreweriesByBreweryId, AddNewUserBrewery } from '../modules/UserBreweryManager';
+import { useParams } from "react-router-dom"
+import { BreweryReviewCard } from './BreweryReviewCard'
+import { BreweryDetailCard } from "./BreweryDetailCard"
+import { UserBreweryCard } from './UserBreweriesCard'
+import { ReviewForm } from '../mybreweries/BreweryReviewForm'
+
+export const BreweryDetail = () => {
+    const [brewery, setBrewery] = useState({})
+    const [usersFrombreweries, setUsersFromBreweries] = useState([])
+    const [isLoading, setIsLoading] = useState(true);
+    const { breweryId } = useParams();
+    const currentUser = parseInt(sessionStorage.getItem("app_user_id"));
+    const getUsersFromUserBreweries = () => {
+        return getUserBreweriesByBreweryId(breweryId)
+        .then(breweriesFromAPI =>  setUsersFromBreweries(breweriesFromAPI))
+    }
+
+    const handleAddToBreweriesToVisit = () => {
+        const newUserBreweryObject = {
+            "userId": currentUser,
+            "breweryId": breweryId,
+            "beenToBrewery": false
+        }
+        AddNewUserBrewery(newUserBreweryObject)
+            .then(() => getUserBreweriesByBreweryId(breweryId).then(setUsersFromBreweries)).then(window.alert("Added to your 'Places I Wanna Go' "))
+    }
+    const handleAddToBreweriesIveBeen = () => {
+        const newUserBreweryObject = {
+            "userId": currentUser,
+            "breweryId": breweryId,
+            "beenToBrewery": true,
+            "review":""
+        }
+        AddNewUserBrewery(newUserBreweryObject)
+            .then(() => getUserBreweriesByBreweryId(breweryId).then(setUsersFromBreweries)).then(window.alert("Added to your 'Places I've Been'"))
+    }
+
+    useEffect(() => {
+        getUsersFromUserBreweries()
+    }, [breweryId])
+
+
+
+    useEffect(() => {
+        getBreweryById(breweryId)
+            .then(b => {
+                setBrewery({
+                    id: b.id,
+                    name: b.name,
+                    state: b.state,
+                    city: b.city,
+                    street: b.street,
+                    website_url: b.website_url,
+                    phone: b.phone,
+                    postal_code: b.postal_code
+                });
+                setIsLoading(false)
+            });
+
+    }, [breweryId]);
+    
+    return (<>
+        <BreweryDetailCard
+            brewery={brewery}
+        />
+                    <UserBreweryCard
+                        brewery={brewery}
+                        // handleAddToBreweriesIveBeen={handleAddToBreweriesIveBeen}
+                        // handleAddToBreweriesToVisit={handleAddToBreweriesToVisit}
+                         />
+               
+        <div>
+            <h3>Other User Reviews:</h3>
+            {usersFrombreweries.filter(user => user.review && user.userId !==currentUser).map(user =>  {
+                    return (
+                        <BreweryReviewCard
+                            key={user.id}
+                            user={user}
+                        />
+                    )
+                } 
+            )}
+            </div>
+
+
+    </>
+    )
+}
