+++ src/components/breweries/BreweryReviewCard.js
@@ -1,11 +1,14 @@
 import React from "react"
-import {Link, useHistory} from 'react-router-dom';
+import { useHistory} from 'react-router-dom';
+
+
+const currentUser = parseInt(sessionStorage.getItem("app_user_id"));
 
 export const BreweryReviewCard = ({user}) =>{
     const history =useHistory();
     return (
         <div className="breweryReviewCard">
-            <h3>User:{user.user.name}</h3>
+            <h4>User:{user.user.name}</h4>
             <p>{user.review}</p>
         </div>
     )
