+++ src/components/mybreweries/MyBreweryCard.js
@@ -20,6 +20,7 @@ export const MyBreweryCard = ({ brewery, handleDeleteMyBrewery, getUserBreweries
                     <p> {brewery.brewery.city}, {brewery.brewery.state}</p>
                     <Link to={`/brewery/${brewery.brewery.id}`}><button className="details">Details</button></Link>
                     <button className="buttonChangeToFalse" type="button" onClick={() => toggleVisited(brewery)}>{brewery.beenToBrewery?"I've not been here yet": "I've been here!"  }</button>
+                    {!(brewery.review = brewery.review) ? "Leave a review on the details page!" : " "}
                     <button className="buttonRemoveMyBrewery" type="button" onClick={() => handleDeleteMyBrewery(brewery.id)}> Remove Brewery  </button>
                 </div>
             </div>
